project/server/requirements.txt
# API
fastapi
uvicorn

#basic
pip
pandas
numpy

# computer vision
opencv-python
pillow

#mysqlconnector
sqlalchemy
mysqlclient
mysql-connector-python